requirements.txt
beautifulsoup4
urllib3
fpdf
Pillow